Solving the problem with a hash table
The problem is solved thanks to the hash table. The program works as follows. Successive numbers from the list are entered into the hash table as keys. Values are volatile, they are assigned as consecutive indexes. Then we check whether the given key does not exist anymore, if it exists, it means that the given number has already appeared. If not, the number is entered into the hash table.