Error explanation
The first function (first solution) has the following problem. As we take numbers sequentially and check whether a given number has not repeated. So if we take the setting [2,5,1,2] then the answer will be 2 which is correct. However, if we take [2,5,5,2], the answer will also be 2, which is incorrect. The second function and the answer resists this drawback and works fine.